+++ Frameworks/FBControlCore.framework/Versions/A/Headers/FBiOSTarget.h
@@ -11,15 +11,17 @@
 
 #import <FBControlCore/FBApplicationCommands.h>
 #import <FBControlCore/FBArchitecture.h>
+#import <FBControlCore/FBBitmapStreamingCommands.h>
 #import <FBControlCore/FBDebugDescribeable.h>
 #import <FBControlCore/FBJSONConversion.h>
 #import <FBControlCore/FBVideoRecordingCommands.h>
 #import <FBControlCore/FBXCTestCommands.h>
 
+@class FBDeviceType;
+@class FBOSVersion;
 @class FBProcessInfo;
+@class FBiOSActionRouter;
 @class FBiOSTargetDiagnostics;
-@protocol FBControlCoreConfiguration_Device;
-@protocol FBControlCoreConfiguration_OS;
 @protocol FBDeviceOperator;
 
 /**
@@ -50,12 +52,12 @@ NS_ASSUME_NONNULL_BEGIN
 /**
  Common Properties of Devices & Simulators.
  */
-@protocol FBiOSTarget <NSObject, FBJSONSerializable, FBDebugDescribeable, FBApplicationCommands, FBVideoRecordingCommands, FBXCTestCommands>
+@protocol FBiOSTarget <NSObject, FBJSONSerializable, FBDebugDescribeable, FBApplicationCommands, FBBitmapStreamingCommands, FBVideoRecordingCommands, FBXCTestCommands>
 
 /**
- Device operator used to control device. It provides API for XCTestBoostrap to interact with the device.
+ A Router for the Reciever.
  */
-@property (nonatomic, nullable, strong, readonly) id<FBDeviceOperator> deviceOperator;
+@property (nonatomic, strong, readonly) FBiOSActionRouter *router;
 
 /**
  The Unique Device Identifier of the iOS Target.
@@ -93,6 +95,16 @@ NS_ASSUME_NONNULL_BEGIN
 @property (nonatomic, copy, readonly) FBArchitecture architecture;
 
 /**
+ The Device Type of the Target.
+ */
+@property (nonatomic, copy, readonly) FBDeviceType *deviceType;
+
+/**
+ The OS Version of the Target.
+ */
+@property (nonatomic, copy, readonly) FBOSVersion *osVersion;
+
+/**
  Process Information about the launchd process of the iOS Target. Currently only applies to Simulators.
  */
 @property (nonatomic, copy, nullable, readonly) FBProcessInfo *launchdProcess;
@@ -101,16 +113,10 @@ NS_ASSUME_NONNULL_BEGIN
  Process Information about the Container Application of the iOS Target. Currently only applies to Simulators.
  */
 @property (nonatomic, copy, nullable, readonly) FBProcessInfo *containerApplication;
-
 /**
- The Configuration of the iOS Target's Device.
+ Device operator used to control device. It provides API for XCTestBoostrap to interact with the device.
  */
-@property (nonatomic, copy, readonly) id<FBControlCoreConfiguration_Device> deviceConfiguration;
-
-/**
- The Configuration of the iOS Target's OS.
- */
-@property (nonatomic, copy, readonly) id<FBControlCoreConfiguration_OS> osConfiguration;
+@property (nonatomic, nullable, strong, readonly) id<FBDeviceOperator> deviceOperator;
 
 /**
  A Comparison Method for `sortedArrayUsingSelector:`