+++ Frameworks/XCTestBootstrap.framework/Versions/A/Headers/FBTestRunnerConfiguration.h
@@ -102,13 +102,4 @@
  */
 - (FBTestRunnerConfiguration *)build;
 
-/**
- @return Default arguments
- */
-+ (NSArray *)defaultBuildAttributes;
-/**
- @return Default environment for a test
- */
-+ (NSDictionary *)defaultBuildEnvironment;
-
 @end