+++ Tests/Integration/Devices/SimulatorTest.m
@@ -3,6 +3,7 @@
 #import "Device.h"
 #import "Simulator.h"
 #import "ShellRunner.h"
+#import "ShellResult.h"
 #import "MachClock.h"
 #import <FBControlCore/FBControlCore.h>
 #import "Application.h"
@@ -23,6 +24,7 @@
 
 - (void)setUp {
     [super setUp];
+    [self quitSimulators];
 }
 
 - (void)tearDown {
@@ -39,22 +41,18 @@
         NSMutableArray *mutable = [NSMutableArray arrayWithCapacity:100];
         for (TestSimulator *simulator in simulators) {
             if (![[simulator stateString] isEqualToString:@"Shutdown"]) {
+                [ShellRunner xcrun:@[@"simctl", @"shutdown", simulator.UDID]
+                           timeout:10];
                 [mutable addObject:simulator];
             }
         }
-
         simulators = [NSArray arrayWithArray:mutable];
-
         return [simulators count] == 0;
     }];
 }
 
 - (void)testBootSimulatorIfNecessarySuccess {
-
-    [self quitSimulators];
-
     NSError *error = nil;
-
     Simulator *simulator = [Simulator withID:defaultSimUDID];
 
     // Boot required
@@ -71,8 +69,6 @@
 }
 
 - (void)testBootSimulatorIfNecessaryFailure {
-    [self quitSimulators];
-
     Simulator *simulator = [Simulator withID:defaultSimUDID];
     FBSimulatorLifecycleCommands *commands;
     commands = [FBSimulatorLifecycleCommands commandsWithSimulator:simulator.fbSimulator];